Separate UsersDiffListSection from UsersDiffList
